Added -u variable to ps1

+++ launch-or-update-local.ps1
@@ -4,17 +4,21 @@
 ###  ./launch-or-update-local.ps1                ###
 ### or without downloading updates...            ###
 ###  ./launch-or-update-local.ps1 -update false  ###
+###  ./launch-or-update-local.ps1 -u false       ###
 ####################################################
 
 param(
         [Parameter()]
-        [ValidateSet('','true','false')]
-        [string]$update
+        [ValidateSet('true','false','')]
+        [string]$update,
+        [Parameter()]
+        [ValidateSet('true','false','')]
+        [string]$u
     )
     
 $chart="src/buyingcatalogue"
 
-if ($update -ne "false")
+if ($update -ne "false" -and $u -ne "false")
 {  
     write-host "Updating Dependencies..."
     Remove-Item $chart/charts/*.tgz
